Chore: refactor engine-api tests to use common create_state_actor function
