+++ engine-api/src/methods/estimate_gas.rs
@@ -60,17 +60,8 @@ async fn inner_execute(
 #[cfg(test)]
 mod tests {
     use {
-        super::*,
-        alloy::primitives::Address,
-        moved::{
-            block::{BlockMemory, Eip1559GasFee, InMemoryBlockQueries, InMemoryBlockRepository},
-            genesis::init_state,
-            primitives::{B256, U256, U64},
-            state_actor::StatePayloadId,
-            storage::InMemoryState,
-        },
-        std::str::FromStr,
-        test_case::test_case,
+        super::*, crate::methods::tests::create_state_actor, alloy::primitives::Address,
+        moved::primitives::U64, std::str::FromStr, test_case::test_case,
     };
 
     #[test_case("0x1")]
@@ -111,26 +102,7 @@ mod tests {
     #[test_case("pending")]
     #[tokio::test]
     async fn test_execute(block: &str) {
-        let genesis_config = moved::genesis::config::GenesisConfig::default();
-        let mut state = InMemoryState::new();
-        init_state(&genesis_config, &mut state);
-
-        let (state_channel, rx) = mpsc::channel(10);
-
-        let state_actor = moved::state_actor::StateActor::new(
-            rx,
-            state,
-            B256::ZERO,
-            genesis_config,
-            StatePayloadId,
-            B256::ZERO,
-            InMemoryBlockRepository::new(),
-            Eip1559GasFee::default(),
-            U256::ZERO,
-            (),
-            InMemoryBlockQueries,
-            BlockMemory::new(),
-        );
+        let (state_actor, state_channel) = create_state_actor();
 
         let state_handle = state_actor.spawn();
         let request: serde_json::Value = serde_json::json!({
